GetData/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.14.3
    # via yfinance
certifi==2026.1.4
    # via
    #   curl-cffi
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   curl-cffi
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.5
    # via google-auth
curl-cffi==0.13.0
    # via yfinance
db-dtypes==1.5.0
    # via -r requirements.in
exchange-calendars==4.13.1
    # via pandas-market-calendars
frozendict==2.4.7
    # via yfinance
google-api-core[grpc]==2.29.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
google-cloud-bigquery==3.40.1
    # via -r requirements.in
google-cloud-bigquery-storage==2.36.1
    # via -r requirements.in
google-cloud-core==2.5.0
    # via google-cloud-bigquery
google-crc32c==1.8.0
    # via google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-bigquery
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   grpcio-status
grpcio-status==1.78.0
    # via google-api-core
idna==3.11
    # via requests
korean-lunar-calendar==0.3.1
    # via exchange-calendars
multitasking==0.0.12
    # via yfinance
numpy==2.4.2
    # via
    #   db-dtypes
    #   exchange-calendars
    #   pandas
    #   yfinance
packaging==26.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
pandas==2.3.3
    # via
    #   -r requirements.in
    #   db-dtypes
    #   exchange-calendars
    #   pandas-market-calendars
    #   yfinance
pandas-market-calendars==5.3.0
    # via -r requirements.in
peewee==3.19.0
    # via yfinance
platformdirs==4.9.2
    # via yfinance
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   yfinance
pyarrow==23.0.1
    # via
    #   -r requirements.in
    #   db-dtypes
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pyluach==2.3.0
    # via exchange-calendars
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2025.2
    # via
    #   pandas
    #   yfinance
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   yfinance
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
toolz==1.1.0
    # via exchange-calendars
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   grpcio
tzdata==2025.3
    # via
    #   exchange-calendars
    #   pandas
urllib3==2.6.3
    # via requests
websockets==16.0
    # via yfinance
yfinance==1.1.0
    # via -r requirements.in
